fix:修复 login 会重复触发的问题

+++ src/utils/http.ts
@@ -38,7 +38,7 @@ export async function Http<R>({
     params.body = JSON.stringify(body);
   }
   return new Promise((resolve, reject) => {
-    fetch('http://8.136.222.10:7001/api' + url, params)
+    fetch('/api' + url, params)
       .then((res) => res.json())
       .then((res) => {
         if (res.status === 200) {
@@ -46,10 +46,12 @@ export async function Http<R>({
           setResult && setResult(res.data);
         } else {
           if (res.status === 1001) {
-            location.hash = `#/login?from=${window.encodeURIComponent(
-              location.hash.slice(1),
-            )}`;
-            localStorage.clear();
+            if (!location.hash.includes('#/login')) {
+              location.hash = `#/login?from=${window.encodeURIComponent(
+                location.hash.slice(1),
+              )}`;
+              localStorage.clear();
+            }
           }
           Toast.fail(res.errMsg);
           reject(res.errMsg);
